Tidy src/compiler

+++ include/compiler/Location.h
@@ -4,11 +4,11 @@ namespace LL2X {
 	class ASTNode;
 
 	struct Location {
-		int line, column, scope;
-		int file = -1;
+		long line, column, scope;
+		long file = -1;
 		/** The index of the location within the debug data section, not within the LLVM IR. */
-		int index = -1;
-		Location(int line_, int column_, int scope_): line(line_), column(column_), scope(scope_) {}
+		long index = -1;
+		Location(long line_, long column_, long scope_): line(line_), column(column_), scope(scope_) {}
 		Location(const ASTNode &);
 	};
 }
